refactor : MainViewModel 에서 마스크스톡리스트 반환하도록 리펙토링
