+++ app/src/main/java/com/survivalcoding/maskinfo/data/model/MaskStockStatus.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.maskinfo.data
+package com.survivalcoding.maskinfo.data.model
 
 enum class MaskStockStatus {
     InSufficient, Spare, Sufficient
